Refactor
- Renames galleryTool.js to ejsImageGallery.js
- Renames m-img-gallery.scss to ejs-img-gallery.scss
