+++ static/css/src/_ejs-img-gallery.scss
@@ -1,5 +1,5 @@
-/* for galleryTool.js */
-.m-img-gallery {
+/* for ejsImageGallery.js */
+.ejs-img-gallery {
     margin-bottom: 1rem;
 
     &__list {
