+++ static/css/src/style.scss
@@ -1,1 +1,1 @@
-@import 'm-img-gallery';
+@import 'ejs-img-gallery';
