+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
     entry: {
-        gallery: './static/js/src/galleryTool.js',
+        gallery: './static/js/src/ejsImageGallery.js',
         index: './static/js/src/index.js',
         withinForm: './static/js/src/withinForm.js',
     },
